Move fuzz targets to own package with Apache headers
