--- poi/src/test/java/org/apache/poi/POIRleFuzzer.java
@@ -1,40 +1,0 @@
-// Copyright 2026 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-////////////////////////////////////////////////////////////////////////////////
-
-package org.apache.poi;
-
-import org.apache.poi.util.RLEDecompressingInputStream;
-
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-
-public class POIRleFuzzer {
-    public static void fuzzerInitialize() {
-    }
-
-    public static void fuzzerTestOneInput(byte[] input) {
-        try (RLEDecompressingInputStream rleStream =
-                     new RLEDecompressingInputStream(new ByteArrayInputStream(input))) {
-
-            byte[] buffer = new byte[1024];
-            while (rleStream.read(buffer) != -1) {
-                // Trigger decompression logic
-            }
-        } catch (IOException | IllegalArgumentException | IllegalStateException | IndexOutOfBoundsException e) {
-            // Expected exceptions on malformed input
-        }
-    }
-}
